making the main menu work again

+++ Source/Game/Application/Configuration/Enums/AutoAimMode.cs
@@ -22,7 +22,7 @@ terms, you may contact me via email at nyvantil@gmail.com.
 */
 
 namespace Game.Application.Configuration.Enums {
-	public enum AutoAimMode : byte {
+	public enum AutoAimMode : uint {
 		Off,
 		Soft,
 		LockOn,
